Rendering performance -- avoid unnecessary DOM renders
In EditorLine (especially), lines were being rendered repeatedly even though the line state
had not changed. Implement a (relatively slow) deep equality for the EditorLine props to
prevent this unnecessary rendering. This improves editor performance tremendously,
especially for operations that did not alter the line state, such as cursor movement.

In addition, implement the same logic in Cursor and DebugEditor.

In TextInput rename the position prop to yPosition to avoid a semantic misunderstanding
with how 'position' is normally used in the project as a char position.

The next performance increase could be obtained by using immutable data structures, and
then shouldComponentUpdate can be implemented with super-fast referene equality
comparisons.